Add merge values to merger buttons and layout fixes
It's super helpful to have the values for each good available at a
glance. This adds the merge values to the merge buttons which (for the
most part) is also the good's value.

There were some breakpoints at which the app was visually quite
broken, the header has changed a bit to acoomodate these sizes.