Use smaller merger buttons
Shrinks the size of the merge buttons and adjusts breakpoints etc
accordingly. This was also an opportunity to optimise the png files
for the company type icons.

Adjusting the breakpoints highlighted that the rice / spice merger
button's text should break lines between the two sections of
text. This is now done as well for great display on screens of any
(reasonable) size.